Split MarkerLayer into seperate layers for markers and nodes.
